refactor: 图片预览脚本抽离公共方法
